Add converters package, use awserrors pkg for all errors
