+++ pkg/cloud/aws/actuators/machine/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
     deps = [
         "//pkg/apis/awsprovider/v1alpha1:go_default_library",
         "//pkg/cloud/aws/services:go_default_library",
-        "//pkg/cloud/aws/services/ec2:go_default_library",
+        "//pkg/cloud/aws/services/awserrors:go_default_library",
         "//pkg/deployer:go_default_library",
         "//pkg/tokens:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
