+++ pkg/cloud/aws/services/awserrors/errors.go
@@ -14,8 +14,19 @@
 package awserrors
 
 import (
+	"net/http"
+
 	"github.com/aws/aws-sdk-go/aws/awserr"
 )
+
+const (
+	AuthFailure        = "AuthFailure"
+	InUseIPAddress     = "InvalidIPAddress.InUse"
+	GroupNotFound      = "InvalidGroup.NotFound"
+	PermissionNotFound = "InvalidPermission.NotFound"
+)
+
+var _ error = &EC2Error{}
 
 // Code returns the AWS error code as a string
 func Code(err error) (string, bool) {
@@ -31,4 +42,99 @@ func Message(err error) string {
 		return awserr.Message()
 	}
 	return ""
+}
+
+// EC2Error is an error exposed to users of this library.
+type EC2Error struct { //nolint
+	err error
+
+	Code int
+}
+
+// Error implements the Error interface.
+func (e *EC2Error) Error() string {
+	return e.err.Error()
+}
+
+// NewNotFound returns a new error which indicates that the resource of the kind and the name was not found.
+func NewNotFound(err error) error {
+	return &EC2Error{
+		err:  err,
+		Code: http.StatusNotFound,
+	}
+}
+
+// NewConflict returns a new error which indicates that the request cannot be processed due to a conflict.
+func NewConflict(err error) error {
+	return &EC2Error{
+		err:  err,
+		Code: http.StatusConflict,
+	}
+}
+
+// NewFailedDependency returns a new error which indicates that a dependency failure status
+func NewFailedDependency(err error) error {
+	return &EC2Error{
+		err:  err,
+		Code: http.StatusFailedDependency,
+	}
+}
+
+// IsFailedDependency checks if the error is pf http.StatusFailedDependency
+func IsFailedDependency(err error) bool {
+	if ReasonForError(err) == http.StatusFailedDependency {
+		return true
+	}
+	return false
+}
+
+// IsNotFound returns true if the error was created by NewNotFound.
+func IsNotFound(err error) bool {
+	if ReasonForError(err) == http.StatusNotFound {
+		return true
+	}
+	return IsInvalidNotFoundError(err)
+}
+
+// IsConflict returns true if the error was created by NewConflict.
+func IsConflict(err error) bool {
+	return ReasonForError(err) == http.StatusConflict
+}
+
+// IsSDKError returns true if the error is of type awserr.Error.
+func IsSDKError(err error) (ok bool) {
+	_, ok = err.(awserr.Error)
+	return
+}
+
+// IsInvalidNotFoundError tests for common aws not found errors
+func IsInvalidNotFoundError(err error) bool {
+	if code, ok := Code(err); ok {
+		switch code {
+		case "InvalidVpcID.NotFound":
+			return true
+		}
+	}
+	return false
+}
+
+// ReasonForError returns the HTTP status for a particular error.
+func ReasonForError(err error) int {
+	switch t := err.(type) {
+	case *EC2Error:
+		return t.Code
+	}
+	return -1
+}
+
+func IsIgnorableSecurityGroupError(err error) error {
+	if code, ok := Code(err); ok {
+		switch code {
+		case GroupNotFound, PermissionNotFound:
+			return nil
+		default:
+			return err
+		}
+	}
+	return nil
 }
