+++ pkg/cloud/aws/services/ec2/BUILD
@@ -8,7 +8,6 @@ go_library(
         "bastion.go",
         "console.go",
         "eips.go",
-        "errors.go",
         "filters.go",
         "gateways.go",
         "instances.go",
@@ -25,11 +24,11 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apis/awsprovider/v1alpha1:go_default_library",
+        "//pkg/cloud/aws/converters:go_default_library",
         "//pkg/cloud/aws/services/awserrors:go_default_library",
         "//pkg/cloud/aws/services/userdata:go_default_library",
         "//pkg/cloud/aws/services/wait:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
-        "//vendor/github.com/aws/aws-sdk-go/aws/awserr:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/service/ec2:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/service/ec2/ec2iface:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
@@ -51,6 +50,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//pkg/apis/awsprovider/v1alpha1:go_default_library",
+        "//pkg/cloud/aws/services/awserrors:go_default_library",
         "//pkg/cloud/aws/services/ec2/mock_ec2iface:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/service/ec2:go_default_library",
